+++ inc/schedule_change.h
@@ -9,11 +9,11 @@ enum class change_type {
 class Schedule_change {
 public:
     int schedule_id;
-    int timestamp;
+    float timestamp;
     change_type type;
     float value;
 
-    Schedule_change(int schedule_id, int timestamp, char type, float value)
+    Schedule_change(int schedule_id, float timestamp, char type, float value)
         : schedule_id(schedule_id), timestamp(timestamp), type(static_cast<change_type>(type)),
           value(value) {}
 };
